test: 로직 변경으로 인한 test 작성
